update create docker net and volume

+++ workspace/automate_scripts/sl_env_setup/create_sldocker_wheel.sh
@@ -21,10 +21,11 @@ do
    esac
 done
 
-sudo docker volume rm sl-cli-lib
+if sudo docker volume list | grep -q 'sl-cli-lib'; then sudo docker volume rm sl-cli-lib; fi
 sudo docker volume create sl-cli-lib
 sudo docker container create --name helper -v sl-cli-lib:/data hello-world
 sudo docker cp -L ./swarmlearning-client-py3-none-manylinux_2_24_x86_64.whl helper:/data
 sudo docker rm helper
 
+if sudo docker network list | grep -q host-"$ip_addr"-net; then sudo docker network rm host-"$ip_addr"-net; fi
 sudo docker network create host-"$ip_addr"-net
